+++ src/tree/prefix_tree_node.h
@@ -7,7 +7,6 @@
 #include <unordered_map>
 #include <map>
 #include <vector> 
-#include "range.h"
 #include "../filter/filter.h"
 
 
@@ -16,7 +15,7 @@
 
 struct node {
     bool is_terminal;
-    std::vector<std::pair<int, node*>> ranges;
+    std::vector<std::pair<uint64_t, node*>> ranges;
     std::vector<filter*> filters;
 
     node();
